Add it wrapper to each test file
